fix(windows): hotkeys offset in config list
Fixes #4426.

After removing the hotkey for the Keyboard Usage pane, the hotkey
`index` value was now offset from the hotkey `Target` value, which
uncovered a bug where the two values were conflated.

Note that the parameter passed to the `keyman:hotkey_set` command is
still called `index` -- this does not need to change.