fix(windows): prevent re-registration of TIPs on 14.0 upgrade
Fixes #4445.

When upgrading from 11.0, 12.0 or 13.0, Keyman needs to re-register the
TIPs so that we use the new registration model for 14.0. However, if we
do this on a 14.0 version install, we end up registering TIPs for every
single suggested language for a keyboard, which is definitely not what
we want.

This fix simply looks to see if we have a 14.0-style installation
already, as evidenced by the presence of a `Transient Language Profiles`
key, and if found, does not continue the 11to13 upgrade process.