fix(web): reset dfltX and dfltY in restorePosition
Fixes #4948.

This resets `keyman.osk.dfltX` and `keyman.osk.dfltY` in the
`restorePosition()` function, except when the user clicks the pin (or
`true` passed in for `keepDefaultPosition` parameter). It also moves the
grungy `window.event` hack out of the function and into a private
handler specifically for the pin div `onclick`.

I didn't want to reset `dfltX` and `dfltY` when the user clicks on the
pin, because that is a change in behaviour.