chore(windows): tweaks to recreatewnd
My initial iteration of `RecreateWnd` function was calling
`HandleNeeded`, which changed the behaviour of `RecreateWnd` in a way
that VCL was not expecting. This caused weird behaviours such as making
the OSK disappear and re-appear.

This version is a little more complicated, because it adds telemetry to
3 functions along with a state flag to avoid unrelated logs, but does
not introduce those side-effects, so it should avoid the problems.