fix(linux): Swap order of dependencies for Debian package
It turns out that the Debian package build system only considers
the first alternative; python3-raven isn't available on the current
Debian version, however we still need it on Ubuntu 18.04 Bionic where
python3-sentry-sdk is not available. This change swaps the
alternatives and so should work both on Debian as well as for
Bionic.