fix(linux): Fix restarting ibus when running with `sudo`
When installing in the shared area the user has to run
`km-package-install` with sudo so that he has sufficient permissions.
In that case we have to do extra work to restart ibus for the
logged-in user instead of root.

Fixes #5309. Partial cherry-pick of #5371.

(cherry picked from commit 0a2dd29e53473b1892e0bc167d850e327814964e)