fix(android): gracefully handle errors in KMW keyboards
Fixes #5397.

This disables the fallback to default keyboard in Keyman Engine for
Android, and prevents a toast from appearing more than once for a given
error. The first five errors will be reported to Sentry and then
subsequent errors will be ignored, until a different keyboard is
selected, to prevent spamming the error log and wasting bandwidth for
the user.