fix(windows): handle restoring modal dialogs consistently
Fixes #5574.

When a user presses Win+M, Win+D, or clicks the desktop button on the
taskbar, all windows get minimized, even those that do not have the
WS_MINIMIZE flag. Delphi modal dialog forms are not expecting this, and
do not restore correctly later.

This scenario arises only when a form is created with a `nil` owner and
then `form.ShowModal` is called.

The fix was already implemented on Splash and HTML forms, but
differently; I have chosen to use the implementation from the Splash
form as it is simpler and seems to work correctly.